+++ l10n/ca.js
@@ -100,12 +100,9 @@ OC.L10N.register(
     "Create form": "Crear formulari",
     "Security": "Seguretat",
     "Light": "Llum",
-    "Classic Light": "Llum clàssica",
     "Dark": "Fosc",
     "Select file to insert text": "Seleccioneu el fitxer per inserir text",
     "Select file": "Seleccioneu un fitxer",
-    "Same as system": "Igual que el sistema",
-    "Contrast dark": "Contrast fosc",
-    "Gray": "Gris"
+    "Same as system": "Igual que el sistema"
 },
 "nplurals=2; plural=(n != 1);");