+++ l10n/da.js
@@ -102,13 +102,10 @@ OC.L10N.register(
     "Run document macros": "Kør dokumentmakroer",
     "Default editor theme": "Standard editortema",
     "Light": "Lys",
-    "Classic Light": "Klassisk lys",
     "Dark": "Mørk",
     "This feature is unavailable due to encryption settings.": "Denne funktion er utilgængelig på grund af krypteringsindstillingerne",
     "Select file to insert text": "Vælg en fil for at indsætte tekst",
     "Select file": "Vælg fil",
-    "Same as system": "Samme som system",
-    "Contrast dark": "Mørk kontrast",
-    "Gray": "Gråt"
+    "Same as system": "Samme som system"
 },
 "nplurals=2; plural=(n > 1);");