+++ l10n/ja.js
@@ -107,7 +107,6 @@ OC.L10N.register(
     "Run document macros": "ドキュメントマクロを実行する",
     "Default editor theme": "エディターのデフォルトテーマ",
     "Light": "明るい",
-    "Classic Light": "ライト(クラシック)",
     "Dark": "ダーク",
     "This feature is unavailable due to encryption settings.": "この機能は暗号化の設定により使用できません。",
     "Enable plugins": "プラグインを有効にする",
@@ -137,8 +136,6 @@ OC.L10N.register(
     "Go to Settings": "設定に進む",
     "Select file to insert text": "テキストを挿入するファイルを選択する",
     "Select file": "ファイルを選択",
-    "Same as system": "システム設定と同じ",
-    "Contrast dark": "ダークコントラスト",
-    "Gray": "灰色"
+    "Same as system": "システム設定と同じ"
 },
 "nplurals=1; plural=0;");