+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## Changed
 - edit xlsb
 - customization_goback setting removed
+- there are only system, light and dark themes
 
 ## 9.9.1
 ## Added